requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
pydantic==2.5.0
aiofiles==23.2.0
python-ffmpeg==2.0.12
celery[redis]==5.3.4
flower==2.0.1